Fixed warnings
git-svn-id: https://svn.code.sf.net/p/snmpb/svn/trunk@196 d4971f4b-c5c8-4039-af2b-e5317b2b9980

+++ app/discovery.cpp
@@ -176,7 +176,7 @@ void DiscoverySnmp::discover(const UdpAddress &start_addr, int num_addr,
 
         if (uaddr.get_ip_version() == Address::version_ipv4)
         {
-            if (iv_snmp_session != INVALID_SOCKET)
+            if (iv_snmp_session != (int)INVALID_SOCKET)
                 sock = iv_snmp_session;
             else
             {
